% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_data.R
\name{balance_data}
\alias{balance_data}
\title{Create New DMUs to Balance Data}
\usage{
balance_data(data, x, y, obs_prop)
}
\arguments{
\item{data}{A \code{data.frame} containing the variables used in the model.}

\item{x}{Column indexes of the input variables in the \code{data}.}

\item{y}{Column indexes of the output variables in the \code{data}.}

\item{obs_prop}{A \code{table} with the observed proportions of efficient and inefficient DMUs.}
}
\value{
It returns a \code{data.frame} with the newly created set of DMUs incorporated.
}
\description{
This function creates new DMUs to address data imbalances. If the majority class is efficient, it generates new inefficient DMUs by worsering the observed units. Conversely, if the majority class is inefficient, it projects inefficient DMUs to the frontier. Finally, a random selection if performed to keep a proportion of 0.65 for the majority class and 0.35 for the minority class.
}
