% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comprobation.R
\name{Comprobation}
\alias{Comprobation}
\title{Mistakes and warnings introduced by user.}
\usage{
Comprobation(method, params, AllMethods, Allparams)
}
\arguments{
\item{method}{models to train defined by user. \code{data}}

\item{params}{params to tune the model defined by user. \code{data}}

\item{AllMethods}{Available models. \code{data}}

\item{Allparams}{Available hyper-parameters. \code{data}}
}
\value{
It returns a \code{data.frame} with the final data.
}
\description{
If user introduced a model not available or not insert the correct hyper-parameters, it will warn to user and it will recommend a solution.
}
