% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeScores.R
\name{compute_scores_additive}
\alias{compute_scores_additive}
\title{Data envelopment analysis' clasification.}
\usage{
compute_scores_additive(data, nX, nY)
}
\arguments{
\item{data}{Dataset \code{data}.}

\item{nX}{Number of inputs \code{data}.}

\item{nY}{Number of outputs \code{data}.}
}
\value{
It returns a \code{data.frame} with the final data.
}
\description{
This function trains for each model, the different hyperparameters and returns the best model with its best hyperparameters.
}
