% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.svm.R
\name{DEA.svm}
\alias{DEA.svm}
\title{Training the different Support Vector Machines.}
\usage{
DEA.svm(data, method_svm, trControl)
}
\arguments{
\item{data}{Dataset \code{data}.}

\item{method_svm}{Different SVM algorithms to train \code{data}.}

\item{trControl}{Training characteristics \code{data}.}
}
\value{
It returns a \code{list} with the chosen model.
}
\description{
This function trains for each model, the different hyperparameters and returns the best model with its best hyperparameters.
}
