% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{optimization}
\alias{optimization}
\title{Optimization problem.}
\usage{
optimization(data, x, y, final_model, orientation)
}
\arguments{
\item{data}{Dataset \code{data}.}

\item{x}{Input value \code{data}.}

\item{y}{Output value \code{data}.}

\item{final_model}{Best trained model \code{data}.}

\item{orientation}{Direction of the optimization. \code{data}.}
}
\value{
It returns a \code{list} with the chosen model.
}
\description{
This function solve the optimization problem and return the solution for each DMU.
}
