% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_data.R
\name{balance_data}
\alias{balance_data}
\title{Create New DMUs to Balance Data}
\usage{
balance_data(
  data,
  x,
  y,
  nX,
  nY,
  number_eff_dmus,
  number_ineff_dmus,
  prop_eff,
  prop_ineff,
  eff_dmus_idx
)
}
\arguments{
\item{data}{A data.frame or matrix containing the variables used in the model.}

\item{x}{Column indexes of the input variables in the \code{data}}

\item{y}{Column indexes of the output variables in the \code{data}}

\item{nX}{Number of inputs \code{data}.}

\item{nY}{Number of outputs \code{data}.}

\item{number_eff_dmus}{Number of efficient DMUs to be created in the \code{data}}

\item{number_ineff_dmus}{Number of inefficient DMUs to be generated in the \code{data}}

\item{prop_eff}{Proportion of efficient DMUs to be introduced in the \code{data}}

\item{prop_ineff}{Proportion of inefficient DMUs to be added to the \code{data}}

\item{eff_dmus_idx}{Index of the efficient dmus observed in \code{data}}
}
\value{
Fill
}
\description{
This function creates new DMUs to address data imbalances. It can generate both efficient and inefficient DMUs.
}
