% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{dea_classification}
\alias{dea_classification}
\title{Training the different machine Learning models.}
\usage{
dea_classification(data, methods, trControl)
}
\arguments{
\item{data}{Dataset \code{data}.}

\item{methods}{Different machine learning algorithms to train \code{data}.}

\item{trControl}{Training characteristics \code{data}.}
}
\value{
It returns a \code{list} with the chosen model.
}
\description{
This function trains for each model, the different hyperparameters and returns the best model with its best hyperparameters.
}
