% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_estimation.R
\name{efficiency_estimation}
\alias{efficiency_estimation}
\title{Efficiency Estimation Using Classification Algorithms}
\usage{
efficiency_estimation(data, x, y, orientation, trControl, methods)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{orientation}{A \code{string}, equal to \code{"input"} (input-oriented) or \code{"output} (output-oriented)}

\item{trControl}{Parameters for controlling the training process (from the \code{'caret'} package).}

\item{methods}{A \code{list} of selected machine learning models and their hyperparameters.}
}
\value{
A \code{"cafee"} object.
}
\description{
This function uses classification algorithms to estimate the efficiency of a set of DMUs (Decision Making Units).
}
