% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_estimation.R
\name{efficiency_estimation}
\alias{efficiency_estimation}
\title{Fill}
\usage{
efficiency_estimation(data, x, y, orientation, trControl, methods)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{orientation}{...}

\item{trControl}{Parameters of the train \code{data}.}

\item{methods}{A list of the chosen ML models' and their hyper-parameters\code{data}.}
}
\value{
Fill
}
\description{
Fill
}
