% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{preprocessing}
\alias{preprocessing}
\title{fill}
\usage{
preprocessing(data, x, y)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}
}
\value{
It returns a \code{matrix} in the required format.
}
\description{
This function arranges the data in the required format and displays error messages.
}
